Print more. Don't unref iochannel since we're saving it. Clean up watch
	* examples/echoserver.c: Print more.
	* examples/echoserver.c (async_accept): Don't unref iochannel
	   since we're saving it.
	* examples/echoserver.c: Clean up watch logic for async
	   server so that there's only a single watch on a socket. (This
	   works around the GLib poll bug.)

	* src/socks-private.h: Fix dport size.

	* src/socks-private.c
(gnet_private_socks_tcp_socket_server_accept):
	   Recreate server socket with same port.  If the programmer
	   specified 0 initially (for an ephemeral port) and cannot
	   get the same port, this fails.

+++ src/socks-private.h
@@ -36,7 +36,7 @@ struct socks5_h {
 	guint8  rsv;
 	guint8  atyp;	
 	guint32 dip;
-	guint32 dport;
+	guint16 dport;
 };
 
 #define GNET_DEFAULT_SOCKS_VERSION 5
